Add files via upload
This project I appended two different tables Bikes and Accessories with the same columns but did not use append queries transformation which appends the secondary query to the current one. The Append Query as NEW. In the AppendQueriesNew, I appended the two tables using the append as new in order to keep both tables intact and created another table which came in as Append1 which was renamed to Products and unloaded the accessories table.